Khoi phuc va xoa vinh vien thuoc tinh

+++ app/Http/Controllers/Admin/AttributeController.php
@@ -57,7 +57,6 @@ class AttributeController extends Controller
         );
 
         if (!$attribute) {
-            // Vietnamese error message
             return redirect()->back()->with(['error' => 'Có lỗi xảy ra, vui lòng thử lại.']);
         }
         foreach ($request->stocks as $key => $value) {
@@ -83,38 +82,83 @@ class AttributeController extends Controller
      */
     public function update(Request $request, Attribute $attribute)
     {
-        $rules = ["attribute_name" => "required", "stocks.*" => "required"];
+        $rules = [
+            "attribute_name" => "required",
+        ];
 
-        foreach ($request->stocks as $key => $value) {
-            $rules["stocks.{$key}.attribute_value"] = 'required';
-            $rules["stocks.{$key}.attribute_quantity"] = 'nullable|numeric';
+        if ($request->stocks) {
+            foreach ($request->stocks as $key => $value) {
+                $rules["stocks.{$key}.attribute_value"] = 'required';
+                $rules["stocks.{$key}.attribute_quantity"] = 'nullable|numeric';
+            }
+        }
+
+        if ($request->old_stocks) {
+            foreach ($request->old_stocks as $key => $value) {
+                $rules["old_stocks.{$key}.value"] = 'required';
+                $rules["old_stocks.{$key}.quantity"] = 'nullable|numeric';
+            }
         }
 
         $messages = [
             'attribute_name.required' => 'Vui lòng nhập tên thuộc tính',
-            'stocks.*.required' => 'Vui lòng nhập giá trị thuộc tính cho sản phẩm',
-            'stocks.*.attribute_value.required' => 'Vui lòng nhập giá trị thuộc tính cho sản phẩm',
+            'stocks.*.attribute_value.required' => 'Vui lòng nhập giá trị thuộc tính',
             'stocks.*.attribute_quantity.numeric' => 'Giá trị không hợp lệ',
+            'old_stocks.*.value.required' => 'Vui lòng nhập giá trị thuộc tính',
+            'old_stocks.*.quantity.numeric' => 'Giá trị không hợp lệ',
         ];
 
         $request->validate($rules, $messages);
 
-        $attribute->update([
-            'name' => $request->attribute_name,
-            'status' => 1
-        ]);
+        $attribute->update(
+            [
+                "name" => $request->attribute_name,
+                "status" => 1
+            ]
+        );
 
         if (!$attribute) {
-            return redirect()->back()->with(['error' => 'Câp nhật thuộc tính không thành công.']);
+            return redirect()->back()->with(['error' => 'Có lỗi xảy ra, vui lòng thử lại.']);
         }
-        foreach ($request->stocks as $key => $value) {
-            AttributeValue::create([
-                'attribute_id' => $attribute->id,
-                'value' => $value['attribute_value'],
-                'quantity' => $value['attribute_quantity'] ?? null,
-            ]);
+
+        $attributeValues = AttributeValue::where('attribute_id', $attribute->id)->get()->toArray();
+        $oldStocks = $request->old_stocks;
+        $newStocks = $request->stocks;
+
+        foreach ($oldStocks as $oldStock) {
+            $oldStock['id'] = array_search($oldStock, $attributeValues);
         }
-        return redirect()->route('admin.attributes.edit', $attribute)->with(['success' => 'Thêm thuộc tính thành công']);
+
+        // So sanh do dai voi so luong phan tu cua mang cu va moi, neu khac nhau thi xoa phan tu cu
+        if (count($attributeValues) != count($oldStocks)) {
+            // so sanh value['id'] cua mang cu va moi, neu khac nhau thi xoa phan tu cu
+            foreach ($attributeValues as $attributeValue) {
+                if (!in_array($attributeValue['id'], array_column($oldStocks, 'id'))) {
+                    AttributeValue::find($attributeValue['id'])->delete();
+                }
+            }
+        } else {
+            foreach ($oldStocks as $oldStock) {
+                $attributeValue = AttributeValue::find($oldStock['id']);
+                $attributeValue->update([
+                    'value' => $oldStock['value'],
+                    'quantity' => $oldStock['quantity'] ?? null,
+                ]);
+            }
+        }
+
+        // Thêm giá trị thuộc tính mới
+        if ($newStocks) {
+            foreach ($newStocks as $newStock) {
+                AttributeValue::create([
+                    'attribute_id' => $attribute->id,
+                    'value' => $newStock['attribute_value'],
+                    'quantity' => $newStock['attribute_quantity'] ?? null,
+                ]);
+            }
+        }
+
+        return redirect()->back()->with(['success' => 'Cập nhật thuộc tính thành công']);
     }
 
     /**
@@ -131,35 +175,36 @@ class AttributeController extends Controller
      */
     public function trashAttribute()
     {
-        $attributes = Attribute::onlyTrashed()->paginate(5);
+        $attributes = Attribute::onlyTrashed()->orderBy('deleted_at', 'desc')->paginate(5);
         return view('admins.attribute.trash', compact('attributes'));
     }
 
     /**
      * Restore the specified resource from storage.
      */
-    public function restoreAttribute(Attribute $attribute)
+    public function restoreAttribute($id)
     {
-        dd($attribute);
-        $attribute->restore();
-        return redirect()->back()->with(['success' => 'Khôi phục thuộc tính thành công']);
+        $attribute = Attribute::withTrashed()->find($id);
+
+        if ($attribute) {
+            $attribute->restore();
+            return redirect()->back()->with('success', 'Khôi phục thuộc tính thành công');
+        }
+
+        return redirect()->back()->with('error', 'Khôi phục thuộc tính thất bại');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function deleteAttribute(Attribute $attribute)
-    {
-        $attribute->forceDelete();
-        return redirect()->back()->with(['success' => 'Xóa thuộc tính vĩnh viễn thành công']);
-    }
+    public function deleteAttribute($id){
+        $attribute = Attribute::withTrashed()->find($id);
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function removeValue(AttributeValue $value)
-    {
-        $value->delete();
-        return redirect()->back()->with(['success' => 'Xóa giá trị thuộc tính thành công']);
+        if ($attribute) {
+            $attribute->forceDelete();
+            return redirect()->back()->with('success', 'Xóa thuộc tính thành công');
+        }
+
+        return redirect()->back()->with('error', 'Xóa thuộc tính thất bại');
     }
 }
